Move SecretsManagerLogo into OSS code
