+++ apps/web/src/app/layouts/secrets-manager-logo.ts
