+++ bitwarden_license/bit-web/src/app/secrets-manager/layout/navigation.component.ts
@@ -4,8 +4,7 @@ import { map } from "rxjs";
 
 import { OrganizationService } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { Organization } from "@bitwarden/common/admin-console/models/domain/organization";
-
-import { SecretsManagerLogo } from "./secrets-manager-logo";
+import { SecretsManagerLogo } from "@bitwarden/web-vault/app/layouts/secrets-manager-logo";
 
 @Component({
   selector: "sm-navigation",
